feat: trying to set up layout grouping and fixing       hydration issues
