+++ app/(index)/loading.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import ProgressBar from '../../components/ProgressBar';
+import Navbar from '../../components/Navbar';
+export default function Loading() {
+  return (
+    <div className='container'>
+      <Navbar />
+      <ProgressBar />
+    </div>
+  );
+}
+
