+++ components/Graph/index.js
@@ -13,55 +13,59 @@ import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import styles from './Graph.module.css';
 import { useEffect } from 'react';
-import dynamic from 'next/dynamic';
+import dynamic from 'next/dynamic'
+
+const DynamicLineGraph = dynamic(() => import('../LineGraph/'), {
+  ssr: false
+})
 
 dayjs.extend(relativeTime);
 
-const LineGraph = ({
-  x,
-  y,
-  label,
-  unit,
-  data,
-  color,
-  domain: [xMin, xMax],
-  props,
-  overlayGraph,
-  dataPoint,
-}) => {
-  const dataRender = [{ id: label, data: formXYSeries(data, x, y) }];
-  const defaultProps = {
-    curve: 'linear',
-    margin: { top: 10, right: 40, bottom: 50, left: 40 },
-    xScale: {
-      type: 'linear',
-      min: xMin,
-      max: xMax,
-    },
-    yScale: {
-      type: 'linear',
-      stacked: false,
-      min: dataPoint.min || 'auto',
-      max: dataPoint.max || 'auto',
-    },
-    enableGridX: false,
-    lineWidth: 1,
-    pointSize: 0,
-    tooltip: (props) =>
-      GraphTooltip({ label, unit, overlayGraph, data, ...props }),
-    axisBottom: {
-      format: (d) => dayjs(d).format('MMM D, YYYY'),
-      tickValues: 3,
-      tickSize: 5,
-      tickPadding: 5,
-      tickRotation: 30,
-    },
-    data: dataRender,
-    colors: [color],
-  };
-
-  return <ResponsiveLineCanvas {...defaultProps} {...props} />;
-};
+// const LineGraph = ({
+//   x,
+//   y,
+//   label,
+//   unit,
+//   data,
+//   color,
+//   domain: [xMin, xMax],
+//   props,
+//   overlayGraph,
+//   dataPoint,
+// }) => {
+//   const dataRender = [{ id: label, data: formXYSeries(data, x, y) }];
+//   const defaultProps = {
+//     curve: 'linear',
+//     margin: { top: 10, right: 40, bottom: 50, left: 40 },
+//     xScale: {
+//       type: 'linear',
+//       min: xMin,
+//       max: xMax,
+//     },
+//     yScale: {
+//       type: 'linear',
+//       stacked: false,
+//       min: dataPoint.min || 'auto',
+//       max: dataPoint.max || 'auto',
+//     },
+//     enableGridX: false,
+//     lineWidth: 1,
+//     pointSize: 0,
+//     tooltip: (props) =>
+//       GraphTooltip({ label, unit, overlayGraph, data, ...props }),
+//     axisBottom: {
+//       format: (d) => dayjs(d).format('MMM D, YYYY'),
+//       tickValues: 3,
+//       tickSize: 5,
+//       tickPadding: 5,
+//       tickRotation: 30,
+//     },
+//     data: dataRender,
+//     colors: [color],
+//   };
+//
+//   return <ResponsiveLineCanvas {...defaultProps} {...props} />;
+// };
 
 const Graph = ({
   graph,
@@ -143,7 +147,7 @@ const Graph = ({
         </div>
       </header>
       <div className={styles.graphWrapper}>
-        <LineGraph
+        <DynamicLineGraph
           {...graph}
           overlayGraph={overlayGraph}
           props={lineGraphProps}
@@ -151,7 +155,7 @@ const Graph = ({
         />
         {overlayGraph && (
           <div className={styles.overlayGraph}>
-            <LineGraph
+            <DynamicLineGraph
               {...overlayGraph}
               props={overlayGraphProps}
               dataPoint={content.dataPoints[overlayGraph.slug]}
